+++ src/components/Login.jsx
@@ -1,91 +1,143 @@
 import { connect } from 'react-redux';
 import { signInAPI } from '../actions';
 import { Redirect } from 'react-router-dom';
-import { useEffect } from 'react';
+import { useState, useEffect } from 'react';
+
+import LoginForm from './LoginComponents/Login';
+import SignUpForm from './LoginComponents/Signup';
+import { ImFacebook2, ImGoogle2, ImTwitter, ImMail3, ImPen, ImWindows8 } from "react-icons/im";
+
 
 import {
   Container,
+  SectionA,
+  SectionB,
+  Form,
   Nav,
-  LoginBtn1,
-  Section,
-  Hero,
+  Close,
+  NavBTN,
 } from '../styles/stylesLogin';
 import { Link } from 'react-router-dom';
 
-const Login = (props) => {
-  useEffect(() => {
-    document.addEventListener('mousemove', parallax);
+// /images/google.svg
 
-    return () => {
-      document.removeEventListener('mousemove', parallax)
+const Login = (props) => {
+  const [login, setLogin] = useState(true);
+  const [show, setShow] = useState(true);
+  const Modal = (props) => {
+    if (!props.show) {
+      return null;
     }
-  });
+
+    return (
+      <Form>
+        <Close onClick={() => setShow(false)}></Close>
+        {login ? (
+          <LoginForm>
+            <span className='sign-up'>
+              Doesn't have an account?{' '}
+              <Link
+                onClick={(e) => {
+                  e.preventDefault();
+                  setLogin(false);
+                }}
+              >
+                Sign up
+              </Link>
+            </span>
+          </LoginForm>
+        ) : (
+          <SignUpForm />
+        )}
+      </Form>
+    );
+  };
 
   return (
-    <Container id='parallax'>
+    <Container>
+      {props.user && <Redirect to='/home' />}
       <Nav>
-        <a href='/' className='logo'>
-          <img src='/images/MaineRRLogo.png' alt='' />
-        </a>
-        <div className='button-container'>
-          <LoginBtn1>
-            <Link to='/agreement'>Login</Link>
-          </LoginBtn1>
-        </div>
-      </Nav>
-
-      <Section>
-        <Hero>
-          <h1>
-            Maine Prisoner Reentry
-            <br />
-            <span>Resources Network</span>{' '}
-          </h1>
-          <div className='assets-container'>
-            <div className='chain' top='270' left='0' speed='3'>
-              <img src='/images/tests/chain.png' alt='' />
-            </div>
-            <div className='birds'>
-              <div className='bird bird-01' speed='6' top='10' left='619'>
-                <img src='/images/tests/bird-1.png' alt='' />
-              </div>
-              <div className='bird bird-02' speed='5' top='65' left='925'>
-                <img src='/images/tests/bird-2.png' alt='' />
-              </div>
-              <div className='bird bird-03' speed='6' top='41' left='744'>
-                <img src='/images/tests/bird-3.png' alt='' />
-              </div>
-              <div className='bird bird-04' speed='5' top='167' left='457'>
-                <img src='/images/tests/bird-4.png' alt='' />
-              </div>
-              <div className='bird bird-05' speed='3' top='200' left='806'>
-                <img src='/images/tests/bird-5.png' alt='' />
-              </div>
-              <div className='bird bird-06' speed='4' top='215' left='724'>
-                <img src='/images/tests/bird-6.png' alt='' />
-              </div>
-              <div className='bird bird-07' speed='4' top='215' left='542'>
-                <img src='/images/tests/bird-7.png' alt='' />
-              </div>
-              <div className='bird bird-08' speed='3' top='294' left='595'>
-                <img src='/images/tests/bird-8.png' alt='' />
-              </div>
-              <div className='bird bird-09' speed='2' top='287' left='508'>
-                <img src='/images/tests/bird-9.png' alt='' />
-              </div>
-              <div className='bird bird-10' speed='1' top='275' left='440'>
-                <img src='/images/tests/bird-10.png' alt='' />
-              </div>
-              <div className='bird bird-11' speed='2' top='318' left='377'>
-                <img src='/images/tests/bird-11.png' alt='' />
-              </div>
-              <div className='bird bird-12' speed='1' top='265' left='377'>
-                <img src='/images/tests/bird-12.png' alt='' />
-              </div>
-            </div>
+        <img src='/images/RRLogo.png' alt='' />
+        <NavBTN
+          className='button'
+          onClick={() => {
+            setLogin(false);
+            setShow(true);
+          }}
+        >
+          <ImPen/>
+          Join Now
+        </NavBTN>
+        <NavBTN
+          className='button'
+          onClick={() => {
+            setLogin(true);
+            setShow(true);
+          }}
+        >
+          <ImMail3/>
+          Email Login
+        </NavBTN>
+        <NavBTN onClick={() => props.signIn()}>
+          <div className='icon'>
+            <ImGoogle2/>
           </div>
-        </Hero>
-      </Section>
+          <span>Google Login</span>
+        </NavBTN>
+        <NavBTN onClick={() => props.signIn()}>
+          <div className='icon'>
+            <ImFacebook2/>
+          </div>
+          <span>Facebook Login</span>
+        </NavBTN>
+        <NavBTN onClick={() => props.signIn()}>
+          <div className='icon'>
+            <ImTwitter/>
+          </div>
+          <span>Twitter Login</span>
+        </NavBTN>
+        <NavBTN onClick={() => props.signIn()}>
+          <div className='icon'>
+            <ImWindows8/>
+          </div>
+          <span>Microsoft Login</span>
+        </NavBTN>
+      </Nav>
+      <SectionA>
+        <Modal show={show} />
+      </SectionA>
+      <SectionB>
+        <div>
+          <h5>
+            Be cautious on social networking sites.
+          </h5>
+            <br />
+          <h5>
+            Remember: once posted, always posted. Protect your reputation on social networks. What you post online stays
+            online. Think twice before posting pictures you wouldn't want your parents or future employers to see.
+          </h5>
+          <br />
+          <h5>
+            Build a positive online reputation.
+          </h5>
+          <br />
+          <h5>
+            Keep personal info personal. Be careful how much personal info you provide on social networking sites.
+            The more information you post, the easier it may be for someone to use that information to steal your
+            identity, access your data, or commit other crimes such as stalking.
+          </h5>
+          <br />
+          <h5>
+            Protect your computer. Security start with protecting your computer. Install Antivirus software. Keep your
+            operating system, web browser, and other software current.
+          </h5>
+          <br />
+          <h5>
+            Use strong passwords. Make sure that your password is at least eight characters long and consists of some
+            combination of letters, numbers, and special characters (for example, +, @, #, or $).
+          </h5>
+        </div>
+      </SectionB>
     </Container>
   );
 };
@@ -101,42 +153,3 @@ const mapDispatchToProps = (dispatch) => ({
 });
 
 export default connect(mapStateToProps, mapDispatchToProps)(Login);
-
-function parallax(e) {
-  const birds = document.getElementsByClassName('bird');
-  const chain = document.getElementsByClassName('chain')[0];
-  moveChain(chain, e);
-  for (let elm of birds) {
-    move(elm, e);
-  }
-}
-
-function move(elm, event) {
-  const speed = elm.getAttribute('speed');
-  const defaultTop = Number(elm.getAttribute('top'));
-  const defaultLeft = Number(elm.getAttribute('left'));
-
-  const x = (window.innerWidth - event.pageX * speed) / 100;
-  const y = (window.innerHeight - event.pageY * speed) / 100;
-
-  const yValue = defaultTop + y;
-  const xValue = defaultLeft + x;
-
-  elm.style.setProperty('top', `${yValue}px`);
-  elm.style.setProperty('left', `${xValue}px`);
-}
-
-function moveChain(elm, event) {
-  const speed = elm.getAttribute('speed');
-  const defaultTop = Number(elm.getAttribute('top'));
-  const defaultLeft = Number(elm.getAttribute('left'));
-
-  const x = (window.innerWidth - event.pageX * speed) / 100;
-  const y = (window.innerHeight - event.pageY * speed) / 100;
-
-  const yValue = defaultTop + y;
-  const xValue = defaultLeft + x;
-
-  elm.style.setProperty('top', `${yValue}px`);
-  elm.style.setProperty('left', `${xValue > 0 ? 0 : xValue}px`);
-}
